+++ anonaddy_shared/src/main/java/host/stjin/anonaddy_shared/managers/SettingsManager.kt
@@ -36,6 +36,7 @@ class SettingsManager(encrypt: Boolean, private val context: Context) {
         NOTIFY_FAILED_DELIVERIES(false, PREFTYPES.BOOLEAN, "notify_failed_deliveries"),
         MANAGE_MULTIPLE_ALIASES(false, PREFTYPES.BOOLEAN, "manage_multiple_aliases"),
         NOTIFY_API_TOKEN_EXPIRY(false, PREFTYPES.BOOLEAN, "notify_api_token_expiry"),
+        NOTIFY_DOMAIN_ERROR(false, PREFTYPES.BOOLEAN, "notify_domain_error"),
         NOTIFY_SUBSCRIPTION_EXPIRY(false, PREFTYPES.BOOLEAN, "notify_subscription_expiry"),
         MAILTO_ACTIVITY_SHOW_SUGGESTIONS(false, PREFTYPES.BOOLEAN, "mailto_activity_show_suggestions"),
 
@@ -81,6 +82,9 @@ class SettingsManager(encrypt: Boolean, private val context: Context) {
         // Used to limit the amount of expiry notifications to 1 a day
         BACKGROUND_SERVICE_CACHE_API_KEY_EXPIRY_LEFT_COUNT(true, PREFTYPES.INT, "cache_api_key_expiry_left_count"),
         BACKGROUND_SERVICE_CACHE_SUBSCRIPTION_EXPIRY_LEFT_COUNT(true, PREFTYPES.INT, "cache_subscription_expiry_left_count"),
+
+        // Used to limit the amount of error notifications to every change
+        BACKGROUND_SERVICE_CACHE_DOMAIN_ERROR_COUNT(true, PREFTYPES.INT, "cache_domain_error_count"),
 
         // This value keeps track of the previous amount of failed deliveries so comparisons can be made in the BackgroundWorker
         BACKGROUND_SERVICE_CACHE_FAILED_DELIVERIES_COUNT_PREVIOUS(true, PREFTYPES.INT, "cache_failed_deliveries_count_previous"),